logging shouldn't swallow this error
> Unless you explicitly raise an error in except block, the Exception is
> caught and stops propagating, and normal execution continues.

- https://twistedmatrix.com/documents/12.0.0/core/howto/defer.html#auto4

`fail` does not raise or return anything, so it swallows the error. Like
not re-raising in an `except` block. This causes normal execution to
resume in callbacks added by callers to this function, which is not what
we want.

Furthermore, we don't need to log anything here at all, since callers
will either log or crash as appropriate. We could log something just in
case a caller fails to handle errors properly, but that would result in
double logging.